Delete obsolete parser and move bottom navigation to its own component
