+++ demo/App.vue
@@ -1,6 +1,6 @@
 <!--
 
-A demo application that lets the user load a YAML serialized surface.
+A demo application that lets the user play with a YAML serialized surface.
 
 -->
 <template>
@@ -13,42 +13,27 @@ A demo application that lets the user load a YAML serialized surface.
         </keep-alive>
       </v-container>
     </v-main>
-    <v-bottom-navigation color="primary" :value="$route.path" @change="(path) => $router.push(path)" fixed app>
-      <v-btn value="/editor">
-        <span>Surface</span>
-        <v-badge :value="error != null" color="error" icon="mdi-error" overlap>
-        <v-icon>mdi-layers</v-icon>
-        </v-badge>
-      </v-btn>
-      <v-btn value="/viewer">
-        <span>Visualization</span>
-        <v-icon>mdi-eye</v-icon>
-      </v-btn>
-      <v-btn value="/svg">
-        <span>SVG</span>
-        <v-icon>mdi-svg</v-icon>
-      </v-btn>
-    </v-bottom-navigation>
     <router-view name="menu" />
+    <bottom-navigation />
   </v-app>
 </template>
 <script lang="ts">
 import { Component, Provide, Vue } from "vue-property-decorator";
 
 import PanZoom from "@/components/PanZoom.vue";
-import Parser from "@/components/Parser.vue";
 import Viewer from "@/components/Viewer.vue";
 import CancellationToken from "@/CancellationToken";
 import Progress from "@/Progress";
 import Overlay from "./Overlay.vue";
 import GlueInteraction from "@/components/interactions/GlueInteraction.vue";
 import PathInteraction from "@/components/interactions/PathInteraction.vue";
+import BottomNavigation from "./BottomNavigation.vue";
 
 @Component({
   components: {
+    BottomNavigation,
     Overlay,
     PanZoom,
-    Parser,
     Viewer,
     GlueInteraction,
     PathInteraction
@@ -72,10 +57,6 @@ export default class App extends Vue {
         // We came here because this process completed. Remove the overlay.
         this.overlay = null;
     }
-  }
-
-  get error() {
-    return this.$store.state.error;
   }
 }
 </script>
